Switch to a single .env file and fix some bugs
This way it's easier to drop in a different docker-compose file for
advanced users. Eventually there should be a UI to do so.

Also this updates to the new syntax of configuring Ultrafeeder.
And it fixes a bug in the way the feeder settings were parsed from the
web form.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>