Add the ability to restore backups
Be careful what is uploaded, try to be careful with the content to avoid
surprises. Ask the user which files they want to restore. And keep the
existing files, just in case (but there is no UI to go back to them, yet
(or ever). Still, via ssh they could be replaced / looked at.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>