implement simplified aggregator selection
The logic looks a bit weird, but it's actually fairly simple. There's
just a bit of boilerplate. Simply assemble the ultrafeeder config with
all aggregators, with only the ones marked as having a privacy policy,
or with the ones that get explicitly set on the second page.

I worry that there are corner cases that I missed, but in my initial
testing this looks right.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>