reorganize the startup process
First we set up the .env file and then pull the Ultrafeeder.
Once that was successful, pull the other containers in the background
and start the tool that should get us the link from my.adsb.im.

Enable all the services the normal way instead of including the link in
the filesystem.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>