redesign the SDR discovery
Simply parse the lsusb output to find the installed SDRs and matching serial
nummbers.

With that in hand, completely redo the setup for multiple SDRs and integrate
handing of the airspy with the RTL-SDRs. This allows for a much smoother user
experience - and avoids a few bugs that were hidden in the old code.

This has been tested with up to four different SDRs (and airspy, one each
RadarBox Flightstick for 978 and 1090, and an RTL-SDR v3).
So far the code assumes that you have up to one SDR for 1090 and up to one for
978. Mostly this is laying the groundwork for other containers also wanting to
access SDRs in the future.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>