Make unchecked checkboxes visible to Flask
By definition, only checked checkboxes get passed to the server. But that
causes us a real problem with the generic update() function, as we need to know
which variables are impacted by a form. Otherwise we can turn settings on, but
not off.

This hack works around this by updating the value for each checkbox (the moment
they are submitted) to 1 or 0, depending on whether they were checked or not,
and then sets all of them to checked. Now Flask gets all of them (since they
were 'checked'), but then uses the value of 0/1 to act on the result - which is
the desired behavior. This doesn't seem sane - but so be it.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>