improve the restore process
The ultrafeeder data wasn't recognized.
If the restore brings changes to the .env or the .yml files in use, we need
to first shut down docker and then bring it up again. We also need to ensure
that we are, in fact, joined to the Zerotier network in the environment.

This will need more testing over time as the scenarios get more complicated,
but it does appear to create working restores between 0.10.x based devices.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>