create generic x86_64 feeder image and run/test in qemu
Create a separate GitHub action for this, currently limited to a test branch.
Create an x86_64 image and copy the kernel into workspace
Run it under qemu with that custom kernel

Give the image some time to boot, then check for the web interface.
Do a basic setup, simulating form input with curl.
Ask the setup system to start the docker containers.
Once that completes, check to see if the advanced setup page and the tar1090
page work.

That should provide at least some reasonable assurance that the image isn't
broken.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>