replace 'None' in env file with empty string
This is a bit annoying - but since I want to allow people to upgrade from older
versions, I need a way to deal with those damaged .env files where we did write
None out as 'None'. This means 'None' can not be a valid value (so, for
example, you cannot name your site 'None').

I think that's acceptable.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>