make random password sed-safe
In an abundance of caution do this right where it is created.
Having a slash in the password confuses our sed expression. Allowing
a backslash (other than to quote a slash) could introduce other
oddities and potentially (if it's the last character) quote the
single quote ending the sed expression, leading to other issues.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>