allow empty value in .env to replace a default value
What a subtle bug. If the value isn't in the .env file, we get None and we should
use the default value. But if the value is in the .env file, but an empty string,
then we should use that instead of the default value.

This caused things to break with an airspy because the default value for the RTL_SDR setting
for Ultrafeeder is 'rtlsdr', and if the user has an airspy that's replaced with '' which
tells the Ultrafeeder readsb not to try to talk to an SDR. But at a restart of the setup app
that was replaced with the default value and now the readsb in Ultrafeeder got confused.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>