Restructure the source layout and move things around
Fundamentally the separation into the adsb-feeder module and the adsb-pi-setup module
no longer makes sense. And the "pi" part of the name makes no sense, anyway.

So this commit makes no(?) functional changes but creates a massive amount of
thrash because it moves so many things around. In the end it simply combines
those two modules into one and adjusts the paths every they showed up in other
parts of the code (e.g. in some of the scripts, the config, etc).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>