limit expert page when not running the full feeder image
This uses the /boot/ADSB-README.txt file as a simple indicator that we are
running a Feeder Image (vs. being installed as an application).
If that is the case, remove the 'system management' part of the expert page as
those function will be handled by standard OS system administration tools.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>