offer different bootstrap when not running the Feeder Image
If this is an app that is installed on a different OS, some of these things
need to be done differently.
We need to manually set an 'image name' (which will just show that this is an
app instead). We also need to define a port other than 80 for the web
interface.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>