create a clear indicator whether this is a feeder image or not
There are some things we need to do differently when we are just running as an
app. And for that we need to be able to definitively know if this is an app or
the full image.

So create a file in the image creation process that otherwise isn't in the file
system. And use that to make the decision at run time. Add a somewhat simple
heuristic that allows the setup app to gues which is which in case the image is
older and we need to figure this out at runtime (as it was installed before the
flag file was included in the build).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>