create some of the DietPi images as install-on-boot
This takes advantage of the new trailing fat partition for these builds.
It allows us to simply setup up the automatic install at boot time, but
instead of installing this as a ADS-B Feeder app on DietPi, we are instead
creating a full ADS-B Feeder Image on the fly - using the git sha from
image build time and the appropriate image names. This also gets you the
full web UI, including the system management features.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>