re-design the feeder update script
Instead of running the script of the version we are coming from, run the
script of the version we are updating to. So first figure out what it is
we are updating to, then force a clean checkout of that, then copy the
new feeder-update script into place, and finally run it with a LAST
parameter of '-go' which tells it that it should just get started.

Adding '-go' at the end of the arguments should allow this to at least
not fail if the user ends up going back to an earlier version. That's
not a well tested thing, not something that should work - but this at
least avoids an obvious error case right at the point of time when this
features gets introduced.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>