ensure GitHub action sees the correct tags
The most recent builds appear to have run based on a tag that was pushed, yet the
checked out code didn't include that tag. My only explanation was that this might
have been some sync delay of the tag across the storage backends that GitHub uses.
This additional refresh of the tags is designed to give us the best possible chance
to create the correct version name in the build.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>