Force requests for aggregator status to use IPv4
Sadly, it appears that we need to switch to using the requests library, which
adds another Python dependency, but at least with that we are able to force
connections to be IPv4.

Since Docker will only connect over IPv4, the feeds will always come from IPv4
addresses. So if the API request to check the feeder status comes from an IPv6
address, that may get incorrect results (observed with TheAirTraffic).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>