make sure the scripts are run as root
With more people running this as an app under a Linux distro, these scripts may
end up being run from the command line instead of just under the control of
adsb-setup. As a result we have seen cases where users run this from their
account and are confused when things don't work.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>