support two stage feeder backend without SDRs
This assumes that there is a second device that has the SDR(s) attached and
provides 'beast_out' on an open port.

Note that this does NOT attempt to also get UAT from a different device. This
is ONLY designed to allow you to get ADS-B / 1090 from a different system.

Fixes #90

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>