open the various ports that people are interested in
I get a request for a different set of ports every few weeks. I don't think it
makes sense to have ten options for ten different things - but if we open them
all if we need any single one, let's just open them all and give up.

Since Docker creates a proxy process for each individual ports, we need to be
specific and can't just open ranges as I tried at first. That would basically
kill the SBC. It worked when I tested it on a rather big and fast VM - but
that's not the target platform.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>