avoid an odd setup issue
A user reports that their system didn't fully come up. The logs indicated that
some network watchdog process on their router kept hitting the feeder homepage
every 30 seconds. As a result it looked like a partial docker start was aborted
and subsequent starts failed as the network bridge was still around from the
partial start. This change tries to do a better job of avoiding multiple
parallel starts, and also adds additional information to the log if a docker
compose start fails.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>