simple DNS liveliness check
Use the Background class to check DNS resolution every hour. This doesn't
actively change the page displayed, instead it will show the error the next
time the home page or basic setup page is shown.

It's most likely to be important when first installing the feeder in a
network with mal-configured DNS configuration.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>