remove rarely used images
So many breaking changes around the Armbian versions disappearing.
Going through the download records, I'll skip all of the SBCs that I can't see
any (or much) interest in. I can always try to bring them back if people
complain, but realistically, 90+% of people use this on an RPi or virtualized
on x86.

The Raspberry32 image is a bit annoying - but Zerotier dropped support for the
older 32bit binaries and I don't want to make a significantly different spin
here. So let's just remove that one as well.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>