add container restart button
Trying to do this from within Python means that we need a seperate thread - and
Flask is making it rather hard to do that without a lot of hoops to jump
through.
So starting a shell script in the background that deals with the task and then
removes a file to indicate that it is done seems much easier.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>