redo the Env storage
Trying to store complex Python types in a file that can also be sourced by a
shell script AND be used as .env file for Docker was just one bridge too many.
So instead use a JSON file as the actual backing store for the Env dict and
simply write out an additional shell/docker style file for those purposes - but
don't include the complex data structures in that.

It would seem to make sense to use the name of the Env variables to determine
which ones are used outside of the Python code, but that clearly isn't the case
so far. So the next commit will create an 'upgrade' tool that allows us to
ingest an existing .env file and that renames things to be more logical.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>