make the first container start after upgrade work
Coming from v1.2.3 or older, before the adsb-setup app is restarted, the .env
file will still include _ADSBIM_STATE_IS_BASE_CONFIG_FINISHED as the variable
name.  After the first time the adsb-setup app runs, this gets converted, but
in order for us not to kill all the running containers, we need to deal with
both of the possible variable names.

Also, don't try to run the inject script if there are no user environment
variables.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>