redo the connection notice for the micro feeder
Show the stage2 systems that have accessed it, or show the likely addresses to
use if not have accessed the micro feeder. This is far more concise and seems
far more useful than the verbose text we had there before. And, as added
benefit, it also gives you the ZT/TS IP addresses to use (if those services are
configured).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>