complete redesign of stage2 data storage
Trying to keep the value based settings for micro feeder, integrated feeder,
and the aggregator instance on a stage2, and having the lists only for the
micro proxies turned the code into an unmanageable mess.

Sometimes you have to cut your losses - so this commit throws all of this
away and changes the structure to follow a much more logical model.

we have four different types of "feeders":
1. integrated feeders (single SBC where one Ultrafeeder collects from SDR and
   send to aggregator)
2. micro feeders (SBC with SDR(s) attached, talking to a stage2 micro proxy)
3. stage2 micro proxies (run on the stage2 system, each talking to a micro
   feeder and to aggregators)
4. stage2 aggregator (showing a combined map of the micro feeders)

Most feeder related values are lists with element 0 being used either for an
integrated feeder, a micro feeder, or the aggregator in a stage2 setup, and
elements 1 .. num_micro_site are used for the micro-proxy instances

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>