rename the misleading constant class and member
These aren't constants. It's the data we share every in the app. So since we
are already rewriting everything, let's straighten this out as well. And while
we are at it, instead of using self._constants hundreds of times, lets go for
better readablity and call it self._d -- so many fewer pointless line wraps.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>