set secure_image to norestore
restoring a backup is the first thing offered when first visiting the
webinterface, this can easily lead to locking yourself out.

As the configuration restore doesn't set an ssh key or root password
either, it doesn't seem appropriate to remember the secure image
configuration.

Let the user explicitely turn on secure image, don't let a backup
restoration enable it (norestore).