allow for multiple feeders to be represented
Any value that is 'per feeder' needs to become a list so that a stage 2 system
can represent many micro feeders.

To keep things sane, for all of these lists index 0 is used for the single
instance on a micro feeder, an integrated feeder, or for the local aggregator
instance on a stage 2 system. And indices 1..N are used for the micro feeder
proxies on a stage 2 system.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>