correctly import v1 config and convert lists
Many config variables changed from being a value of a certain type to being a
list of that value, with item 0 of that list being used for a simple feeder
setup. That's a bit wasteful and complex, but keeps things nice and consistent
in the code.

There are other instances where the type change needs to be explicitly handled,
e.g. where we now have int values instead of strings with the number in them.

Finally - the idea to switch from the rather narrow "MLAT_SITE_NAME" to
"SITE_NAME" was well intentioned, but makes upgrades from really v1.2 and v1.3
feeders even more complicated, so instead of adding more special handling code,
let's just go back to the old name. This is all fragile enough as it is.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>