first attempt to direct user to stage 2 micro feeder map
This solution feels crude, but it seems to work.
If there were any changes to the settings, the Ultrafeeder will likely take a
moment to start up. We somehow need to wait long enough before redirecting
until the map is ready. This tries to check if the map is up before redirecting
the user in order to give a better user experience - but it only tries for five
times before moving forward anyway.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>