stage2: allow simple reuse of the history and graph data
Using numbered directories is a bit more painful when deleting a micro feeder
in the middle - and creates a real issue when adding it again because then the
history is lost.
By using the IP address for the intermediate folder name, we make it easy to
keep the data around.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>