stage2: improve the micro feeder connection code
Instead of checking the state of the feeder from the user's browser, do it from
the actual stage2 system - and be much smarter about figuring out if this is a
very recent adsb.im (with micro_settings api), v1.2.3 or newer (with at least
the base_info) or something else (where we check if it at least gives us
beast).

In the latter case there is still a lot more work needed to get to a better
user experience - but step by step this is getting better.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>