invert default mlat privacy setting and improve explanation
In a number of conversations I ended up convinced that the better overall
outcome would be to default this setting to off (so sites are visible at
approximate locations) and make it more explicit that this is the default. If
the user doesn't really care, we should err on the site of creating a better
experience for the community - and that is better served if more mlat sites are
shown (at approximate locations) on those maps if more mlat sites are shown (at
approximate locations) on those maps.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>