add multi outline for stage 2
If you are running a stage 2 setup with multiple micro feeders, this creates an
'intelligent' outline that combines overlapping areas, but keeps separate areas
separate.
So if you have multiple feeders in the vicinity of each other, this creates a
larger combined actual outline (instead of a weird radial outline around a
single fixed point).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>