Build DietPi images that are pre-installed
We still run all of the installs again at boot time, but since
they all already ran when the DietPi image was built, the are
moslty no-ops (with the exception of enabling and starting some
of the services which doesn't 'stick' from the preinstall).

This still should significantly improve the firstboot time.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>