create various VM images on GitHub
Now that we are prying apart the DietPi build we can use the freshly added
--createvms option to take the raw image and continue processing; this way we
get ready to go VM images instead of just a .vmdk.

Unfortunately that makes the 'running out of space' issue even worse. So let's
try to clean  up even earlier and see if that allows us to make it all the way
through.

This also takes a step at simplifying the naming and making the upload logic
more consistent.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>