avoid crashes in multioutline
If a range outline received from a micro feeder isn't a simple Linear Ring
(which can happen even in fairly benign cases when we first start up, and
commonly happens if the defined site location is wrong and all the planes seen
are in an area that doesn't contain the site location), the Polygon code would
fail. This carefully works around this.

Also, simplify the code that keeps track of how many polygons there are.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>