allow editing of micro feeder name and IP
It's possible that the person running the stage 2 cannot edit the micro feeder
name but wants to use a different name to tell the sites apart. So allow them
to edit the name.

In a fairly different scenario it's plausible that the IP address of a micro
feeder changes (most likely if the feeder is connected locally and gets its IP
address via DHCP - remove feeders via Tailscale or Zerotier should have
consistent addresses).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>