stage2: don't ignore the micro feeder name if it's unset so far
If we create a new micro feeder proxy from an adsb.im feeder, but decide not to
import the settings (which means we aren't copying the data sharing settings,
etc), we will still grab the base info from the micro feeder. Therefore in the
case of do_import == False there's still the case where we need to store the
site name we just detected.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>