data sharing: docker compose if "request key" is used in UI
If the user requests a key on the data sharing page and that request is
successful, the key is saved to the configuration but the settings
aren't applied until the user clicks 'i'm done here'.

Fix that by setting seen_go whenever a key is requested.
When saving values to the configuration, those values should also be
applied via docker compose.