Yet another redo of the top level menu layout
Move live map to a button near the top left of the screen and then group all
the maps and map options under Map in the main menu.

Rename System Mgmt: Since it's under the System top level entry, calling it
Management should be sufficient and it avoids the abbrv.

Thanks @ottergoose

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>