avoid emptry listeners in the list
In order to specify the type of an Env we use the default value. Which in the
case of arrays ends up setting the value for the first item. In most cases
that's fine, because the position in the array has semantic meaning, so when an
actual value is written this gets simply replaced.  But in the case where an
array simply collects items (like for the stage2 IP addresses that have
accessed a micro feeder), this created the odd situation where the first stage2
to have accessed the micro feeder was always the empty string.

This commit both prevents this for new installations and hides the silly empty
strings for existing ones.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>