stage2: ensure site names are unique
The UI code relies on the fact that the micro feeder site names and the stage2
site name all are unique. It breaks in really strange ways if they aren't. So
let's just enforce that.

Yes, the appended underscore(s) are really ugly, but this is at least a simple
deterministic process to ensure that they are all different.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>