stage2: update micro feeder information via javascript
This has two major impacts. One is the obvious one - as the user stares at the
stage2 page, within 5 minutes an update of the feeder to a newer version will
be reflected on this page. The other one is even nicer. This significantly
speeds up the initial load of the stage2 page for people with many micro
feeders. With ten micro feeders across VPN tunnels it sometimes took more than
15s to load the stage2 page.  Now it's near instantaneous.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>