stage2: clean up multi-outline and multi-heywhatsthat code
There were a number of ways this could fail for odd outlines.
This adds tests to make sure the heywhatsthat outlines actually exist,
and a number of additional validity checks. It also retains the json
files created to simplify debugging in the future - no point to create
a tmp dir every time we call this code.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>