hotspot: change workflow to be less confusing
After some testing, it seems like the attempt to come back to hotspot mode to
give the user the good news that the connection to the provided wifi was
successful is confusing and really redundant.
Come back to the hotspot if the WiFi connection failed, but otherwise simply
connect to the WiFi and move on - the user will figure out that the hotspot
went away.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>