waiting app: show the log while waiting
This modifies the waiting app to auto scroll a log file while waiting to be
done with the current task.

This can be used as a placeholder website while doing a feeder upgrade and even
during the initial boot in order to give the user a response when they access
the feeder with a web browser.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>