stage2: add UI option to reduce bandwith beteen micro feeder and stage2
This is on my default because it drops no positions and no other relevant messages.
The one visible 'downside' (as it is) is a reduced number of messages shown on the stage2
screens, but nothing of value is skipped.
This only works with new enough micro feeders that have a new enough Ultrafeeder to provide
that data on port 30006.
What's still missing is an option to turn this on/off for already established microfeeders.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>