add reasonable range defaults for tar1090
A stage2 could get data from anywhere on the globe. But an integrated feeder or
micro feeder really shouldn't be able to see planes more than 300nm away.
Hopefully this will cut down on some of the outliers we see from GPS spoofing
and from occasional glitches when pilots turn on the GPS (apparently some GPS
receivers love to claim that they are in Kansas for a moment...)

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>