Use better property names to distinguish capability vs use
The terms used made it easy to get confused whether a micro feeder was capable
of doing beast reduce optimize for mlat or whether it actually WAS using that
protocol.

While we are added, add the same for being capable of sending UAT data (i.e.,
the micro feeder actually has a UAT radio).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>