setup mDNS alias for site name when running as image
This way the user can access the feeder / stage2 by its name instead of simply
'adsb-feeder.local'.

Of course, the name may have to be adjusted to be a valid hostname.
Since the avahi publish service seems to have issues with '-' (which would
otherwise be allowable in a hostname), we restrict the name to only the
alphanumeric characters in the site name.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>