DietPi builds: use pre-built images
It was incredibly wasteful to always recreate these images. Instead use release
builds for the adsbim branch of DietPi.

Also, it's silly to run the clone / move of the feeder image code at first
boot. The build process carefully ensures that the correct version of the
feeder image software is already in place. All the dependencies are provided
by the adsbim branch of DietPi.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>