include GPS derived lat/lng/alt in base_info
We do nothing to actually enable and ingest the GPS data into the feeder,
that's all up to the user to set up. But if they do, then this shows the
GPS location to a connected stage2.

This is a first step in some very minimal support to make mobile micro
feeders possible.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>