ensure adsb-feeder.local still works and allow '-' in hostname
Previously we had disallowed '-' in hostnames because that seemed to be
unsupported in avahi-publish. Turns out the bug was in the systemd unit file.
%I replaces non-alphanumeric values in the argument with '/' -- using %i fixes
that problem. So now we can actively publish adsb-feeder.local, but can also
include dashes in hostnames. To avoid things from going poorly for people who
upgrade, we ALSO publish a .local name that has all the dashes removed.

Also update release notes.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>