add get_temperatures.json api endpoint
This simply gets the values from a file and returns an empty object if that
file doesn't exist. The implementation of how to populate this file is missing,
and nothing currently consumes this API.

The expected format of that json file is {"cpu":"xx","ext":"yy","now":"zzzz"}
where xx is the CPU temperature in deg C, yy is the temperature of an external
temperature sensor (typically either the temperature of the enclosure of the
SDR) and zzzz is the time stamp in seconds since epoch.

github.com/adsbim/adsb-temperature-sensing is a very rough example of how a
service could create this json file from the data obtained from a DHT22
attached to the SBC expansion port.

The API endpoint simply forwards that json file, adding an "age" field that
provides the age of the measurement in seconds (which is then used by the
javascript in base.html to change the opacity of the temperature information as
it gets older).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>