switch to the logs being implemented through journal
This is a fairly painful conversation, but the reasons why this was a
file really no longer make a lot of sense and make other things harder.
So let's convert things.

The one downside is that I haven't figured out a good way to have a
manual start of the feeder-update also be logged as a unit.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>