add support for local SDR on stage2 system
This is a fairly invasive change that ends up special-casing some situation.

In principle this should be simple: we use a local nanofeeder to deal with the
SDR and automatically add a stage2-proxy for that local nanofeeder.

In reality there are weird port conflicts and we need to make sure that the
stage2 combined Ultrafeeder doesn't just use the SDR directly.

So as one side effect this resulted in having to copy the docker-compose.yml
into a stage2-main.yml which explicitly removes the ability to talk to a local
SDR.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>