beast/mlat status: add ultrafeeder based aggregator status analysis
This changes the status type to reflect more situations:
- connected / disconnected / unknown (as before)
- bad (timeouts for mlat)
- warning (small number of timeouts for mlat but majority good data
which resulted in some UI changes, including the introduction of new symbols
for the different status values.

One of the positive side effects is that we can stop using docker calls to get
to the mlat data for adsbx.

Generally we assume that if there is API data for beast, we prefer that over
what we can get from Ultrafeeder. But we assume that the mlat-client data that
we get from Ultrafeeder is likely more accurate than what we can get from an
aggregator's API, so we prefer that.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>