skeleton for ambient temperature tracking
The code in temperature.py doesn't work without the sensor specific
infrastructure being present - which seems like it would be a bit too much to
deal with in the feeder image. So that's left for the user to experiment with.

Two sample implementations (with further external dependencies) for the RPi and
OPi are provided for inspiration.

All this does is the structure how it works for me on my boards.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>