stage2: implement runtime checks for micro feeder lists
This way we'll get a warning if we somehow managed to lose consistency across
the different micro feeder related variables.

We need to temporarily disable these warnings any time we add or remove a micro
feeder. To ensure all is well we trigger a consistency check after each of
those operations.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>