stage2: add number of planes to stats
This way a third party tool (like a HomeAssistant integration) can show the
number of planes currently received / daily total for each of the feeders.

Using sets for this makes it easy to collect the unique planes that the feeder
has seen. To avoid pointless writes, this uses a signal handler to write out
the json on exit. That makes the startup code a bit less elegant, but it seems
worth that price.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>