statistics: show chart title with end date
This way we can simply use the day number in the legend of the chart and avoid
the month/day -- day/month confusion. Also this allows an easy way to point out
that the dates are UTC based.

Also removes a couple of verbose debugging outputs that snuck into the last
beta.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>