fr24: add dns_search to yml
this prevents a rare connection issue mostly when people run pihole DNS

with a search domain set in /etc/resolv.conf typicall from DHCP, the search
domain is added by docker daemon DNS proxying which is a bug.

this results in query responses such as :: which in turn seem to make
fr24feed use that response and thus not being able to connect.
another bug.

in other words the perfect storm