handle a non-stage2 feeder with no data source
Having a system that isn't a stage 2 and that doesn't have either a local SDR
that is configured, or a remote SDR that it gets data from seems a bit
pointless. So let's not report that as 'unusually low data' but instead as
'not working'. And point the user to the SDR setup page to fix this.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>