webUI: improve horizontal scrolling of feeding table
When scrolling the table with sticky first column, the content of the table is
visible in a small strip to the left of the first column as it slides past
'below' that first column. With this change there is still a tiny, one pixel
wide area that shows the artifact, but it's far less pronounced.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>