store key SDR properties in the SDR object
In order for this to work, we cannot recreate the SDRs every time we scan
USB for devices. Instead we create a persistant data set accessed by SDR
serial number that simply gets assigned to the list of currently found SDRs.

This first pass only remembers the purpose of an SDR in its object.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>