correctly handle SDR with usage that's disabled
When the user disables one of the non-ADS-B usages and then edits the SDR
that was assigned to that usage, the SDR needs to be switched to 'other' in
order for the UI to make sense.

Also use the 'changed' event handler to set the gain and biastee fields visible
(or not) - that way we don't need the same code in two different places.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>