Fix hotspot POST handling
The hotspot app starts a DNS server that resolves all names to itself in
order to create the captive portal. It also starts a Flask server to
serve the hotspot web UI, which has a catch-all endpoint that shows the
captive portal no matter what the client requests.

But that endpoint used to assume that any POST request must be a set of
wifi credentials it should try and would shut down the hotspot to try
them whenever it received one. This would also get triggered on
completely unrelated requests, e.g. POST requests to whatsapp.com/chat
(which the app apparently likes to make). In effect, background activity
out of the user's control can trigger seemingly random hotspot
shutdowns.

To prevent this, only try credentials if the request actually contains
the required keys. Otherwise, render the hotspot page.