correctly handle 'other' assignment for SDRs
The new UI changes resulted in 'other' status being handled incorrectly. In
order to store this state correctly, it must be 'numbered' so we don't get
confused by multiple SDRs that were marked 'other' - this fixes that error
without confusing the user (as they still just see 'other' as usage).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>