add capability to change the serial number of an RTLSDR
This has been a frequently requested feature, but it's tricky to get right and
it's unclear how many weird corner cases this might still be missing.

While the rtl_eeprom command works with multiple SDRs connected, there's no
obvious way to tell which entry in its device list corresponds to which SDR
that is plugged into the system. So to simplify that part of the problem the UI
insists that all RTLSDRs except for the one that the user wants to modify are
removed.

Once that has happened, all containers that might be accessing SDRs are stopped
and the new serial is written to the SDR. After which we run into the next
challenge. While on some boards (including a Raspberry Pi 4) it may be possible
to disable power to a USB port and through this simulate an unplug/plug
sequence, the only reliable way to ensure that the correct serial is used
appears to be to physically unplug and replug the SDR - which of course the UI
cannot really observe. As a result the UI flow for this process is not as good
as one might hope.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>