streamline the way we deal with purposes and their Env vars
The sdr,py code shouldn't know about Env variables. And having the .purposes()
method return the tags for the Env vars was extremely confusing.

So here we have the purposes in exactly one place (sdr.py) and then do the rest
of the translations in the app code. And in the process add a couple of snazzy
helper functions that make the code more consistent.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>