deal with automatic assignments to 'other'
When an SDR gets replaced for its purpose, the UI code doesn't know which
other-<i> to assign to it, so we need to fix this on the app side.
Additionally, the order of SDRs might have changed after probing, so be extra
sure to remove a re-assigned serial from any 'other-*' values it may have been
assigned to before.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>