Make button size and styling reactive to screen size
Depending on screen width, let's use the smaller button style everywhere.

This adds two somewhat arbitrary break points, simply based on trial and error.
Not sure those are ideal - especially as they aren't aligned with the other
size based changes made elsewhere...

The commit also tries to clean up the order of code as I made a bit of a mess
of that in an earlier commit.

Fixes: #294

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>